fix implicit uint8_t -> double narrowing
YES, the standard committee believes this might be narrowing
and clang 12 throws this error:

> error: non-constant-expression cannot be narrowed from type 'uint8_t' (aka 'unsigned char') to 'double' in initializer list [-Wc++11-narrowing]

+++ plugins/ros1_parsers/fiveai_stamped_diagnostic.h
@@ -100,7 +100,7 @@ class FiveAiDiagnosticMsg : public RosMessageParser
       {
         auto key = fmt::format("{}/{}/status", _topic_name, replaced_key);
         auto& series = getSeries(key);
-        series.pushBack({ timestamp, diag.status });
+        series.pushBack({ timestamp, static_cast<double>(diag.status) });
       }
 
     }
